at31.2

+++ src/app/home/lead/lead.component.ts
@@ -1,11 +1,10 @@
 import { Component, OnInit, Inject } from '@angular/core';
-import { TaskService } from "../../services/task.service";
+import { TaskService } from '../../services/task.service';
 import { Router } from '@angular/router';
-import {SelectionModel} from '@angular/cdk/collections';
+import { SelectionModel } from '@angular/cdk/collections';
 import { MatTableDataSource } from '@angular/material/table';
-import {MatDialog} from '@angular/material/dialog';
+import { MatDialog } from '@angular/material/dialog';
 import { CadastroComponent } from './cadastrolead/cadastro.component';
-
 
 export interface LeadComponent {
   nome: string;
@@ -17,24 +16,33 @@ export interface LeadComponent {
 @Component({
   selector: 'app-lead',
   templateUrl: './lead.component.html',
-  styleUrls: ['./lead.component.scss']
+  styleUrls: ['./lead.component.scss'],
 })
 export class LeadComponent implements OnInit {
-
-  displayedColumns: string[] = ['email', 'nome', 'telefone', 'cidade', 'encerrar', 'editar', 'deletar'];
-  projects:any = [];
+  displayedColumns: string[] = [
+    'email',
+    'nome',
+    'telefone',
+    'cidade',
+    'encerrar',
+    'editar',
+    'deletar',
+  ];
+  projects: any = [];
   dataSource = new MatTableDataSource<any>();
   selection = new SelectionModel<LeadComponent>(true, []);
   isUpdated: any;
   _id: any;
 
-
-
-  constructor(private taskService: TaskService, private router: Router, public dialog: MatDialog) { }
+  constructor(
+    private taskService: TaskService,
+    private router: Router,
+    public dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
-      this.taskService.getContato().subscribe(
-        (res: { projects: any; }) => {
+    this.taskService.getContato().subscribe(
+      (res: { projects: any }) => {
         this.dataSource.data = res.projects;
         this.projects = res.projects;
       },
@@ -42,48 +50,52 @@ export class LeadComponent implements OnInit {
     );
   }
 
-  applyFilter($event: Event){
+  applyFilter($event: Event) {
     const filterValue = ($event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
   deletar(item: any) {
-    this.taskService.deleteLead(item._id).subscribe((res) => {
-      window.location.reload();
-      // this.router.navigate(['/lead']);
-    })
+    let text = `Deseja deletar esse lead?`
+    if (confirm(text) == true)
+    {
+      this.taskService.deleteLead(item._id).subscribe((res) => {
+        window.location.reload();
+      });
+    } else {
+      
+    }
+
   }
 
-  home(){
+  home() {
     this.router.navigate(['/home']);
   }
 
-  relatorios(){
+  relatorios() {
     this.router.navigate(['/formulario']);
   }
 
-  cadlogin(){
+  cadlogin() {
     this.router.navigate(['/cadlogin']);
   }
 
-  addcurso(){
+  addcurso() {
     this.router.navigate(['/addcurso']);
   }
 
-  lead(){
+  lead() {
     this.router.navigate(['/lead']);
   }
 
-  vendas(){
+  vendas() {
     this.router.navigate(['/vendas']);
   }
 
   openDialog(project: any, isUpdated = false) {
     const dialogRef = this.dialog.open(CadastroComponent, {
       width: 'auto',
-      data: {...project, isUpdated}
+      data: { ...project, isUpdated },
     });
   }
 }
-
-
