+++ src/app/home/vendas/vendas.component.ts
@@ -1,12 +1,17 @@
 import { Component, OnInit } from '@angular/core';
-import { TaskService } from "../../services/task.service";
-import { Router } from "@angular/router";
-import { animate, state, style, transition, trigger } from '@angular/animations';
+import { TaskService } from '../../services/task.service';
+import { Router } from '@angular/router';
+import {
+  animate,
+  state,
+  style,
+  transition,
+  trigger,
+} from '@angular/animations';
 import { MatTableDataSource } from '@angular/material/table';
 import { SelectionModel } from '@angular/cdk/collections';
 import { MatDialog } from '@angular/material/dialog';
 import { EditvendasComponent } from './editvendas/editvendas.component';
-
 
 export interface PeriodicElement {
   curso: string;
@@ -22,24 +27,35 @@ export interface PeriodicElement {
   styleUrls: ['./vendas.component.scss'],
   animations: [
     trigger('detailExpand', [
-      state('collapsed', style({height: '0px', minHeight: '0'})),
-      state('expanded', style({height: '*'})),
-      transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')),
+      state('collapsed', style({ height: '0px', minHeight: '0' })),
+      state('expanded', style({ height: '*' })),
+      transition(
+        'expanded <=> collapsed',
+        animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')
+      ),
     ]),
   ],
-
 })
 export class VendasComponent implements OnInit {
-
-  displayedColumns: string[] = ['curso', 'grau', 'duracao', 'valor', 'descricao'];
-  projects:any = [];
+  displayedColumns: string[] = [
+    'curso',
+    'grau',
+    'duracao',
+    'valor',
+    'descricao',
+  ];
+  projects: any = [];
   dataSource = new MatTableDataSource<any>();
   clickedRows = new Set<PeriodicElement>();
   selection = new SelectionModel<VendasComponent>(true, []);
   expandedElement!: PeriodicElement;
   isUpdated: any;
 
-  constructor(private taskService: TaskService, private router: Router, public dialog: MatDialog) {}
+  constructor(
+    private taskService: TaskService,
+    private router: Router,
+    public dialog: MatDialog
+  ) {}
 
   ngOnInit() {
     this.taskService.getTasks().subscribe(
@@ -48,7 +64,6 @@ export class VendasComponent implements OnInit {
         this.projects = res.projects;
       },
       (err) => console.log(err)
-
     );
   }
 
@@ -59,12 +74,17 @@ export class VendasComponent implements OnInit {
   }
 
   deletar(item: any) {
-    this.taskService.deleteProject(item._id).subscribe((res) => {
-      this.router.navigate(['/vendas']);
-    })
+    let text = "Deseja deletar essa venda?"
+    if (confirm(text) == true)
+    {
+      this.taskService.deleteProject(item._id).subscribe((res) => {
+        this.router.navigate(['/vendas']);
+      });
+    } else {
+    }
   }
 
-  applyFilter($event: Event){
+  applyFilter($event: Event) {
     const filterValue = ($event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
     this.dataSource.filter = filterValue.trim().toUpperCase();
@@ -73,8 +93,7 @@ export class VendasComponent implements OnInit {
   openDialog(project: any, isUpdated = false) {
     const dialogRef = this.dialog.open(EditvendasComponent, {
       width: '250px',
-      data: {...project, isUpdated}
+      data: { ...project, isUpdated },
     });
   }
 }
-
